+++ icekit/articles/abstract_models.py
@@ -1,12 +1,6 @@
-from icekit.abstract_models import FluentFieldsMixin
-from icekit.publishing.models import PublishingModel
+from icekit.publishing.models import PublishableFluentContents
 from django.db import models
 
-# TODO: this should ideally be in icekit.abstract_models, but doing so
-# creates circular import errors.
-class PublishableFluentModel(FluentFieldsMixin, PublishingModel):
-    class Meta:
-        abstract = True
 
 class TitleSlugMixin(models.Model):
     # TODO: this should perhaps become part of a wider ICEkit mixin that covers
@@ -22,7 +16,7 @@ class TitleSlugMixin(models.Model):
         return self.title
 
 
-class PublishableArticle(PublishableFluentModel, TitleSlugMixin):
+class PublishableArticle(PublishableFluentContents, TitleSlugMixin):
     '''
     Basic Article type (ie that forms the basis of independent collections of
     publishable things).
@@ -30,4 +24,3 @@ class PublishableArticle(PublishableFluentModel, TitleSlugMixin):
 
     class Meta:
         abstract = True
-